+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Non-standard operator `<>` can be used as an alternative for `!=` 
   to avoid ambiguity with a verbatim block mark (`!`) inside `if` blocks.
+- Shortcut attribute values `.CLS` and `#ID` no longer need to be surrounded 
+  by spaces, so `div#main.left.top` is now correct.
+
 
 ## [1.1.4] - 2021-07-20